@@ -3,8 +3,9 @@
   "version": "0.11.0",
   "description": "viz.js the JavaScript Library with API support for VIZ blockchain",
   "main": "lib/index.js",
+  "type": "module",
   "scripts": {
-    "test": "mocha -t 20000 --require babel-polyfill --require babel-register",
+    "test": "mocha -t 20000",
     "test-auth": "npm test -- --grep 'viz.auth'",
     "build": "npm run build-browser && npm run build-node",
     "build-browser": "rm -rf dist && NODE_ENV=production webpack && gzip -k -f ./dist/*.js && du -h ./dist/*",
@@ -35,49 +36,53 @@
   },
   "homepage": "https://github.com/VIZ-Blockchain/viz-js-lib#readme",
   "dependencies": {
-    "babel-preset-env": "^1.7.0",
     "bigi": "^1.4.2",
-    "bluebird": "^3.5.5",
-    "browserify-aes": "^1.0.6",
-    "bs58": "^4.0.0",
-    "buffer": "^5.0.6",
+    "bluebird": "^3.7.2",
+    "browserify-aes": "^1.2.0",
+    "bs58": "^6.0.0",
+    "buffer": "^6.0.3",
     "bytebuffer": "^5.0.1",
-    "create-hash": "^1.1.2",
-    "create-hmac": "^1.1.4",
-    "cross-env": "^5.0.0",
-    "cross-fetch": "^3.0.0",
-    "debug": "^2.6.8",
-    "detect-node": "^2.0.3",
-    "ecurve": "^1.0.5",
-    "lodash": "^4.16.4",
-    "secure-random": "^1.1.1",
-    "ws": "^1.1.1"
+    "create-hash": "^1.2.0",
+    "create-hmac": "^1.1.7",
+    "cross-env": "^10.1.0",
+    "cross-fetch": "^4.1.0",
+    "debug": "^4.4.3",
+    "detect-node": "^2.1.0",
+    "ecurve": "^1.0.6",
+    "lodash": "^4.17.21",
+    "secure-random": "^1.1.2",
+    "ws": "^8.18.3"
   },
   "devDependencies": {
-    "babel-cli": "^6.16.0",
-    "babel-eslint": "^7.1.1",
-    "babel-loader": "^6.2.5",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "^6.16.0",
-    "babel-preset-es2017": "^6.16.0",
-    "babel-register": "^6.14.0",
-    "eslint": "^3.5.0",
-    "eslint-plugin-import": "^1.15.0",
-    "eslint-plugin-jsx-a11y": "^2.2.2",
-    "eslint-plugin-react": "^6.2.1",
-    "json-loader": "^0.5.4",
-    "mocha": "^5.2.0",
+    "@babel/cli": "^7.28.3",
+    "@babel/core": "^7.28.5",
+    "@babel/preset-env": "^7.28.5",
+    "@babel/register": "^7.28.3",
+    "@eslint/js": "^9.39.2",
+    "assert": "^2.1.0",
+    "babel-loader": "^10.0.0",
+    "crypto-browserify": "^3.12.1",
+    "eslint": "^9.39.2",
+    "eslint-plugin-import": "^2.32.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-react": "^7.37.5",
+    "mocha": "^11.7.5",
     "mocha-make-stub": "^2.3.2",
-    "should": "^11.1.0",
-    "uglifyjs-webpack-plugin": "^1.3.0",
-    "webpack": "^1.13.2",
-    "webpack-visualizer-plugin": "^0.1.5"
+    "process": "^0.11.10",
+    "should": "^13.2.3",
+    "stream-browserify": "^3.0.0",
+    "terser-webpack-plugin": "^5.3.16",
+    "util": "^0.12.5",
+    "webpack": "^5.104.1",
+    "webpack-bundle-analyzer": "^5.1.0",
+    "webpack-cli": "^6.0.1"
   },
   "contributors": [
     "Fabien (https://github.com/bonustrack)",
     "James Calfee (https://github.com/jcalfee)",
     "Nilesh Suthar (https://github.com/nil1511)",
     "Pedro Tacla Yamada (https://github.com/yamadapc)",
-    "Anatoly Piskunov (https://github.com/on1x)"
+    "Anatoly Piskunov (https://github.com/on1x)",
+    "Vladimir Babin (https://github.com/chiliec)"
   ]
 }
