@@ -1,9 +1,13 @@
-'use strict';
-const Visualizer = require('webpack-visualizer-plugin');
-const UglifyJSPlugin = require('uglifyjs-webpack-plugin')
-const _ = require('lodash');
-const path = require('path');
-const webpack = require('webpack');
+import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
+import TerserPlugin from 'terser-webpack-plugin';
+import _ from 'lodash';
+import path from 'path';
+import webpack from 'webpack';
+import { fileURLToPath } from 'url';
+import process from 'process';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
 const DEFAULTS = {
   isDevelopment: process.env.NODE_ENV !== 'production',
@@ -11,89 +15,68 @@ const DEFAULTS = {
 };
 
 function makePlugins(options) {
-  const isDevelopment = options.isDevelopment;
-
-  let plugins = [
-    new Visualizer({
-      filename: './statistics.html'
-    }),
-  ];
-
-  if (!isDevelopment) {
-    plugins = plugins.concat([
-      new webpack.optimize.DedupePlugin(),
-      new UglifyJSPlugin({uglifyOptions: { ...options }}),
-      new webpack.optimize.AggressiveMergingPlugin(),
-    ]);
+  const plugins = [];
+  if (!options.isDevelopment) {
+    plugins.push(
+      new webpack.ProvidePlugin({
+        Buffer: ['buffer', 'Buffer'],
+        process: 'process/browser',
+      }),
+      new BundleAnalyzerPlugin({
+        analyzerMode: 'static',
+        openAnalyzer: false,
+        reportFilename: 'statistics.html',
+      })
+    );
   }
-
   return plugins;
 }
 
-function makeStyleLoaders(options) {
-  if (options.isDevelopment) {
-    return [
-      {
-        test: /\.s[ac]ss$/,
-        loaders: [
-          'style',
-          'css?sourceMap',
-          'autoprefixer-loader?browsers=last 2 version',
-          'sass?sourceMap&sourceMapContents',
-        ],
-      },
-    ];
-  }
-
-  return [
-    {
-      test: /\.s[ac]ss$/,
-      loader: ExtractTextPlugin.extract(
-        'style-loader',
-        'css!autoprefixer-loader?browsers=last 2 version!sass'
-      ),
-    },
-  ];
-}
-
-function makeConfig(options) {
-  if (!options) options = {};
+function makeConfig(options = {}) {
   _.defaults(options, DEFAULTS);
-
   const isDevelopment = options.isDevelopment;
 
   return {
-    devtool: isDevelopment ? 'cheap-eval-source-map' : 'source-map',
+    mode: isDevelopment ? 'development' : 'production',
+    devtool: isDevelopment ? 'eval-cheap-module-source-map' : 'source-map',
     entry: {
       viz: path.join(options.baseDir, 'src/browser.js'),
-      'viz-tests': path.join(options.baseDir, 'test/api.test.js'),
     },
     output: {
       path: path.join(options.baseDir, 'dist'),
       filename: '[name].min.js',
+      clean: true,
     },
     plugins: makePlugins(options),
     module: {
-      loaders: [
-        {
-          test: /\.js?$/,
-          exclude: /node_modules/,
-          loader: 'babel',
-        },
-        {
-          test: /\.json?$/,
-          loader: 'json',
-        },
+      rules: [],
+    },
+    optimization: {
+      minimize: !isDevelopment,
+      usedExports: true,
+      sideEffects: true,
+      mergeDuplicateChunks: true,
+      minimizer: [
+        new TerserPlugin({
+          terserOptions: {
+            compress: true,
+            mangle: true,
+          },
+        }),
       ],
+    },
+    resolve: {
+      extensions: ['.js'],
+      fallback: {
+        assert: 'assert/',
+        stream: 'stream-browserify',
+        buffer: 'buffer/',
+        util: 'util/',
+        crypto: 'crypto-browserify',
+      },
     },
   };
 }
 
-if (!module.parent) {
-  console.log(makeConfig({
-    isDevelopment: process.env.NODE_ENV !== 'production',
-  }));
-}
-
-exports = module.exports = makeConfig;
-exports.DEFAULTS = DEFAULTS;
+export default makeConfig;
+export { DEFAULTS };
